feat(Lavanderia) adicionado ServicoDao com insert e selectAll

+++ src/modulo-05-java/aula02/jdbc/src/br/com/cwi/crescer/model/Cliente.java
@@ -11,7 +11,7 @@ public class Cliente {
         return idCliente;
     }
 
-    public void setIdCliente(int idCliente) {
+    public void setIdCliente(long idCliente) {
         this.idCliente = idCliente;
     }
